added the api endpoint route for signing up users
